+++ tests/test_testing.py
@@ -48,3 +48,22 @@ from pyairtable import testing as T
 def test_fake_function(funcname, sig, expected):
     func = getattr(T, funcname)
     assert func(*sig.args, **sig.kwargs) == expected
+
+
+def test_coerce_fake_record():
+    assert T.coerce_fake_record({"Name": "Alice"}) == {
+        "id": ANY,
+        "createdTime": ANY,
+        "fields": {"Name": "Alice"},
+    }
+    assert T.coerce_fake_record({"fields": {"Name": "Alice"}}) == {
+        "id": ANY,
+        "createdTime": ANY,
+        "fields": {"Name": "Alice"},
+    }
+    assert T.coerce_fake_record({"id": "rec123", "fields": {"Name": "Alice"}}) == {
+        "id": "rec123",
+        "createdTime": ANY,
+        "fields": {"Name": "Alice"},
+    }
+    assert T.coerce_fake_record(fake := T.fake_record()) == fake